refactor: move Encrypter to criptography folder
