+++ src/data/usecases/add-account/db-add-account-protocols.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
 export * from '../../../domain/usecases/add-account'
 export * from '../../../domain/models/account'
-export * from '../../protocols/encrypter'
+export * from '../../protocols/criptography/encrypter'
 export * from '../../protocols/add-account-repository'
